move neighbor_to_arrow into raster_funcs
