+++ src/landlab/components/flow_director/flow_director_d8.py
@@ -260,73 +260,3 @@ class FlowDirectorD8(_FlowDirectorToOne):
         self._grid["node"]["flow__sink_flag"][sink] = True
 
         return receiver
-
-
-def neighbor_to_arrow(x, diagonals_ok=True):
-    """Convert an array of neighbor nodes to flow direction arrows.
-
-    Parameters
-    ----------
-    x : array_like
-        Indices to neighbor elements.
-    diagonals_ok : bool, optional
-        Allow diagonal neighbors.
-
-    Returns
-    -------
-    ndarray
-        Array for directions represented by arrows that point to
-        neighbor elements.
-
-    Examples
-    --------
-    >>> from landlab.components.flow_director.flow_director_d8 import neighbor_to_arrow
-    >>> x = [
-    ...     [0, 0, 1, 2],
-    ...     [0, 0, 3, 10],
-    ...     [4, 8, 9, 10],
-    ... ]
-
-    The *○* symbol means an element points to itself.
-
-    >>> neighbor_to_arrow(x)
-    array([['○', '←', '←', '←'],
-           ['↑', '↖', '↗', '↙'],
-           ['↑', '←', '←', '←']], dtype='<U1')
-
-    Set the `diagonals_ok` keyword to ``False`` to exclude diagonal neighbors.
-    The default it to include diagonal neighbors.
-
-    >>> neighbor_to_arrow(x, diagonals_ok=False)
-    array([['○', '←', '←', '←'],
-           ['↑', '?', '?', '?'],
-           ['↑', '←', '←', '←']], dtype='<U1')
-    """
-    import numpy as np
-
-    x = np.asarray(x)
-
-    n_rows, n_cols = x.shape
-
-    d4_arrows = {
-        (-1, 0): "↑",
-        (1, 0): "↓",
-        (0, -1): "←",
-        (0, 1): "→",
-        (0, 0): "○",
-    }
-    d8_arrows = d4_arrows | {
-        (-1, -1): "↖",
-        (-1, 1): "↗",
-        (1, -1): "↙",
-        (1, 1): "↘",
-    }
-    direction_arrows = d8_arrows if diagonals_ok else d4_arrows
-
-    shift_row = x // n_cols - np.arange(n_rows).reshape((-1, 1))
-    shift_col = x % n_cols - np.arange(n_cols)
-
-    def get_arrow(row_shift, col_shift):
-        return direction_arrows.get((row_shift, col_shift), "?")
-
-    return np.vectorize(get_arrow)(shift_row, shift_col)
